Stomp.Frames

Description

The Frames module provides abstract data types for representing STOMP Frames, and convenience functions for working with those data types.

Documentation

data Header

A Header contains a name and a value for a STOMP header

Constructors

Header HeaderName HeaderValue

data Headers

A Headers is a recursive, list-like data structure representing the set of Headers for a given STOMP frame

Constructors

Some Header Headers

EndOfHeaders

data Body

A Body represents the body of a STOMP frame. It is either empty or consists of a single ByteString.

Constructors

EmptyBody

Body ByteString

data Command

A Command represents the command portion of a STOMP Frame.

Constructors

SEND

SUBSCRIBE

UNSUBSCRIBE

BEGIN

COMMIT

ABORT

ACK

NACK

DISCONNECT

CONNECT

STOMP

CONNECTED

MESSAGE

RECEIPT

ERROR

data Frame

An abstract data type representing a STOMP Frame. It consists of a Command, Headers, and Body.

Constructors

Frame Command Headers Body

data AckType

Data type representing the possible acknowledgement types for a subscription to a STOMP broker.

Constructors

Auto

Client

ClientIndividual

addFrameHeaderFront :: Header -> Frame -> Frame

Given a Frame, add a Header to the front of its Headers.

addHeaderEnd :: Header -> Headers -> Headers

Add a Header to the end of the Headers

addHeaderFront :: Header -> Headers -> Headers

Add a Header to the front of the Headers

addReceiptHeader :: String -> Frame -> Frame

Add a receipt Header to the Frame.

connect :: String -> Frame

Generate a CONNECT Frame given a host identifier as a String.

connected :: String -> Frame

Generate a CONNECTED Frame given a version identifier as a String.

disconnect :: String -> Frame

Generate a DISCONNECT Frame given a receipt identifier as a String.

errorFrame :: String -> Frame

Generate an ERROR Frame given an error message as a String.

getBody :: Frame -> Body

Convenience function for retrieving the Body portion of a Frame.

getCommand :: Frame -> Command

Convenience function for retrieiving the Command portion of a Frame.

getContentLength :: Headers -> Maybe Int

Given Headers, get the value of the content-length Header if it is present.

getDestination :: Frame -> Maybe String

Given a Frame, get the value of the destination header if it is present.

getHeaders :: Frame -> Headers

Convenience function for retrieving the Headers portion of a Frame.

getId :: Frame -> Maybe String

Given a Frame, get the value of the id header if it is present.

getReceipt :: Frame -> Maybe String

Given a Frame, get the value of the receipt header if it is present.

getReceiptId :: Frame -> Maybe String

Given a Frame, get the value of the receipt-id header if it is present.

getSupportedVersions :: Frame -> Maybe [String]

Given a Frame, get a list of supported STOMP versions, provided that the accept-version Header is present and well-formed.

getValueForHeader :: String -> Headers -> Maybe String

Given a header name and a Frame, get the value for that header if it is present.

makeHeaders :: [Header] -> Headers

Given a list of Header datatypes, return a Headers datatype.

messageIdHeader :: String -> Header

Given a message identifier as a String, generate a message-id Header.

receipt :: String -> Frame

Generate a RECEIPT Frame given a receipt identifier as a String.

sendText :: String -> String -> Frame

Generate a plain text SEND Frame given a message as a String and a destination as a String.

subscribe :: String -> String -> AckType -> Frame

Generate a SUBSCRIBE Frame given a subscription identifer and destination as Strings, an

subscriptionHeader :: String -> Header

Given a subscription identifier as a String, generate a subscription Header.