Stomp.Frames.IO

Description

The IO module of the Frames package encapsulates error-handling IO operations on Handles that are expected to be receiving STOMP frames.

Documentation

data FrameHandler

A FrameHandler encapsulates the work of sending and receiving frames on a Handle. In most cases, this will be a Handle to a TCP socket connection in a STOMP client or broker.

data FrameEvt

A FrameEvt is a type of event that can be received from a FrameHandler (see the get function).

Constructors

NewFrame Frame

ParseError String

GotEof

initFrameHandler :: Handle -> IO FrameHandler

Given a resource Handle to which STOMP frames will be read from and written to, initializes a FrameHandler and returns it to the caller.

put :: FrameHandler -> Frame -> IO ()

Puts the given Frame into the given FrameHandler. This function will block until the Frame has been processed.

get :: FrameHandler -> IO FrameEvt

Get the next FrameEvt from the given FrameHandler. This function will block until a FrameEvt is available.

close :: FrameHandler -> IO ()

Closes the Handle and kills all threads associated with the FrameHandler.

frameToBytes :: Frame -> ByteString

Convert a Frame to a STOMP protocol adherent ByteString suitable for transmission over a handle.