Stomp.Frames.Router

Description

The Router module implements an asynchronous notification system for events received on a FrameHandle. Callers can request various types of events and receive those events on a dedicated communications channel.

Documentation

initFrameRouter :: FrameHandler -> IO RequestHandler

Given a FrameHandler on which FramesEvts are being received, initalize a FrameRouter and return a RequestHandler for that FrameRouter to the caller.

requestResponseEvents :: RequestHandler -> IO (SChan FrameEvt)

Given a RequestHandler, request a new SChan on which to receive response events. Response events are defined as any Frame received other than a MESSAGE Frame.

requestSubscriptionEvents :: RequestHandler -> String -> IO (SChan FrameEvt)

Given a RequestHandler, request a new SChan on which to receive subscription events for a given subscription identifier.

data RequestHandler

The RequestHandler is used to request various types of notifications from the FrameRouter.