Stomp.Increment

Description

The Increment module provides a stateful incrementer.

Documentation

data Incrementer

An Incrementer maintains the state of an Integer that increases by one with each access.

newIncrementer :: IO Incrementer

Initialize a new incrementer and return it in an IO context. The first value retrieved will be 0.

getNext :: Incrementer -> IO Integer

Get the next sequential Integer and return it in an IO context.

getNextEvt :: Incrementer -> Evt Integer

Get the next sequential Integer and return it in an Evt context.

getLast :: Incrementer -> IO Integer

Get the most recent value returned by the Incrementer and return it in an IO context.

getLastEvt :: Incrementer -> Evt Integer

Get the most recent value returned by Incrementer and return it in an Evt context.