Stomp.TLogger

Description

The TLogger module implements a transactional logger. It is a convenient way to write output to a handle in a multi-threaded context in which it is necessary to provide sequential locking access to the handle. Examples of situations in which this might be used is in a logfile or a command-prompt in a multi-threaded application.

Documentation

initLogger :: Handle -> IO Logger

Initialize and return a logger for the given Handle.

dateTimeLogger :: Handle -> IO Logger

Initialize and return a logger that automatically timestamps all of its messages with the current UTC time.

log :: Logger -> String -> IO ()

Send a line of output to the Logger's Handle.

prompt :: Logger -> String -> IO ()

Send output to the Logger's handle, but do not append a newline.

addTransform :: (IO String -> IO String) -> Logger -> Logger

Get a new Logger for the same Handle as the original Logger, but with some additional String transformation applied.

data Logger

A Logger can be used to send output to its Handle. It is guaranteed to be thread-safe.