Stomp.Util

Description

The Util module exports convenience functions from the Stomp library.

Documentation

tokenize :: String -> String -> [String]

Given a delimiter and a String, return a list of the tokens in the String given by splitting on that delimiter.